feat: Adds background process with storage
This update adds in the background process which adds Chrome listeners to oversee all network
traffic and tab activity. The background process also persists this data in storage, and performs
cleanup when tabs are closed. This will allow the background process to act like a server and update
the Popup HTML and other UI elements with data as if it was a server.