refactor: Replace "url-parser" with "minimatch"
Minimatch is a far more supported and used library for simple glob matching. While not specific to
URLs, it has essentially the exact same functionality and is in fact more forgiving. In addition,
its formatting will be familiar to NodeJS developers, as it's the internal matching library used by
NPM.