refactor: Removes abstraction of tab data before App component
It no longer made sense to only pass down some Tab data, as tons of properties were being used and
the risk of forgetting to pass them down was too high. This project would make for an ideal
TypeScript codebase, considering the plethora of data properties that have to be passed in every
direction.