feat: Re-adds error for no tab data & adds custom DL backend
With the transition of the Popup app to the new stream connection model, it lost the ability to
receive errors from the Background. This resulted in it hanging when it had no data available. This
re-adds errors to the Background and Popup. In addition, this adds the skeleton of the Custom Data
Layer feature. The API for Chrom Storage is added in and test values are saved, retrieved, and
injected onto the Window object for retrieval by the Injected Script. This will allow the Injected
Script to pull these values and merge it with the values from the Config.