refactor: Requests updated Data Layers at the BG script-level
While requesting an update from the tab upon clicking the Popup technically worked, it had the
downside that someone refreshing an existing Full Report would not receive the updated content,
since it would not retrigger the polling. This pings the Content Script whenever a new connection is
made. Since the Content Script checks for the script actively running, it isn't wasteful.